+++ src/pong/PongGame.cpp
@@ -14,10 +14,6 @@ void PongGame::InitGame(const Color clearColor){
     // Load the game scene using SceneManager
     _SceneManager->LoadScene<PongGameScene>();
     _GameScene = dynamic_cast<PongGameScene*>(_SceneManager->GetCurrentScene());
-    
-    // Set the next scene (game over scene) that will be loaded when game scene completes
-    if (_GameScene)
-        _GameScene->SetNextScene<PongGameOverScene>(_GameScene);
 }
 
 PongGameScene* PongGame::GetGameScene() const